+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- `_BenchmarkOnIsolate`:
+  - Added optional field `shutdownIsolateDelay`.
+  - Ensure that the `Isolate` closes its `receivePort`.
+
 ## 1.0.0
 
 - Initial version.